add simple release management with `cargo-release`
Added a `[release]` section to the Cargo.toml file to establish a common standard for managing release settings. This new addition incorporates settings for generating changelogs, signing commits, and updating dependencies. This will ensure a more streamlined and uniform release procedure. It also corrects the script to conform to the targeted behavior when releasing alpha versions.